fixed decimal serailizing

+++ src/main/java/Payment/Client/SerializationConfiguration/DecimalSerializer.java
@@ -5,12 +5,11 @@ import com.fasterxml.jackson.databind.JsonSerializer;
 import com.fasterxml.jackson.databind.SerializerProvider;
 
 import java.io.IOException;
-import java.math.BigDecimal;
 
-public class DecimalSerializer extends JsonSerializer<BigDecimal> {
+public class DecimalSerializer extends JsonSerializer<Double> {
 
     @Override
-    public void serialize(BigDecimal bigDecimal, JsonGenerator jsonGenerator, SerializerProvider serializerProvider) throws IOException {
+    public void serialize(Double bigDecimal, JsonGenerator jsonGenerator, SerializerProvider serializerProvider) throws IOException {
         jsonGenerator.writeString(bigDecimal.toString());
     }
 }
